requirements/dev.txt
-r ./base.txt


# DJANGO
# ----------------------------------------------------------------
django-debug-toolbar==3.2.1
django-extensions==3.1.3
ipython==7.25
autopep8==1.5.7